Reorder shapes (with GUI), snap to grid, bug fixes and refactoring
- Moved color information and key information to be stored in CanvasArea
in order to limit global variables.
- Refactored shape selection and drawing of selected shape into separate
method. Removed "selected" attribute from individual shapes. So now
instead of telling a shape that it is selected and haiving to manually
toggle them on and off as needed, there is just one call on the current
shape's drawSelected() mehtod. This allowed me to clean up the logic for
selection and helped fix some bugs.
- Embedded grid menu inside menuButton in order to clean up GUI.
- Created GUI for handling reordering of shapes.
- Added method to swap elements in the shape array.
- Wrote method to round mouse x and y to grid spacing in order to snap
to grid when drawing or moving shape.
- Changed finishShape method for each shape to also set the handle
radius so that we have a base radius and rotation to reset to.