Canvas Settings and Window sizing
- Added feature to adjust size of canvas. User can input new size into
text field and the canvas will be changed and scaled accordingly. Also
adjusts the gridDensity to match new canvas.
- Created custom menuButton to contain GUI elements for the above.
- Added a couple preferences to controls panel and stage to handle
sizing. Set minimum on both to prevent squishing and set it so that the
minimum window size is the size of both bars.
- Added tool tip for arrow buttons to clarify their use.
- Added mouse coordinates (for debugging when trying to adjust canvas)
but I may keep them as a toggleable feature.
- Usability: when last feature is deleted you now automatically switch
from edit mode to draw mode as there is nothing to edit.