Copy, snap, coordinates
- Created functionality in GUI to create a copy of current shape and add
it to list of shapes. Used a copy constructor and copy method in each
class to reference the current attributes of the shape and make a new
shape that is an exact copy of the current shape.
- Updated menu to control mouse coordinates. Can now be set to display
at the mouse, on the top of the canvas, or set to off.
- Updated menu to control grid. Grid on and grid snap can now be toggled
inside the menu and are shortcutted to ctrl + g and ctrl + l. Menu is
synced to other GUI elements so that they are consistent.
- Repaired gridSnap to work with lines and polygons.