Exporting and Background
- Created methods for each shape to print a string output of the
processing code required to generate the shape. The method collectes the
strings from each shape and copies them to the clipboard as either
independent shapes or as a processing "file" (ie the void setup and void
draw methods are included).
- Created methods for each shape to add their geometry to graphic object
and return it. Then the main program can save this graphic object to a
file.
- Added export options to menu.
- Added GUI element to control background color. (TODO: move to another
location to separate it from other similar controls. )