IO Operations, Reference Image, SVG output
- Added Option to save as SVG. Uses functionality built into PGraphics
to generate a SVG.
- Added functionaliy and GUI element to handle loading reference image.
When importing image user will be promted if reference will not fit
within canvas and user can opt to resize the canvas or just allow the
ref image to be cut off. There is also a toggle for the reference image
to be turned on or off.
- Created method to clear screen.
- Updated Circle class and handle class with a save method. This
collects all of the information about the object and stores it as a
string.
- Implemented a load constructor in Circle and Handle class. This
constructor takes in a string array (the string generated from the save
method) and uses it to create a new shape.
- The "open" option in the main menu loads a previous file and the save
function creates a save file.