Scaling handles, usability - control + z, load options
- Added method to scale handles based on scaleFactor in order to avoid
large handles on small canvases and tiny handles on large canvases.
TODO: The math still needs some tweaking.
- Added canvas inforamtion to save file so that the file loads the
canvas the way that it was when saved.
- Created system to handle control + z (undoing changes).  Implemented
Change class to hold information about the changes the user is making
and store the information about the previous state inside Change
objects.
- Implemented a stack data structure to hold the most recent 20 Changes
(can adjust in code) . When a new Change past 20 changes occurs it will
simply overwrite the oldest change. When the user presses ctrl + z, the
top item in the stack is popped off and removed. Then the undo function
is called on that Change object to revert the most recent user change.